Refs #117. refactor to move stitching-related code to stitch_dos.py
